+++ src/error.rs
@@ -1,4 +1,3 @@
-use crate::ProjectCompileOutput;
 use semver::Version;
 use std::{
     io,
@@ -68,9 +67,6 @@ pub enum SolcError {
 
     #[error("no artifact found for `{}:{}`", .0.display(), .1)]
     ArtifactNotFound(PathBuf, String),
-
-    #[error("{0}")]
-    CompilationFailed(ProjectCompileOutput),
 
     #[cfg(feature = "project-util")]
     #[error(transparent)]
